feat: Move publishing container and release
Move the build and publish steps into the publish workflow
Any change that GitHub Actions makes doesn't trigger a new workflow. So
we need to grab the new version from the tag_version step an then
publish using that.

We now publish containers on the pull_request step for testing.

Stop publishing the sha as a tag, no one uses this.

Signed-off-by: Dan Webb <dan.webb@damacus.io>